+++ async/_util.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-export function _exponentialBackoffWithJitter(
+export function exponentialBackoffWithJitter(
   cap: number,
   base: number,
   attempt: number,
